# Conflicts:
#	src/main/java/com/sysc4806/project/controllers/IndexController.java
#	src/main/java/com/sysc4806/project/controllers/ProductController.java
#	src/main/java/com/sysc4806/project/controllers/UserController.java
#	src/main/java/com/sysc4806/project/models/Product.java
#	src/main/resources/templates/index.html
#	src/main/resources/templates/productSearch.html
#	src/main/resources/templates/products.html
#	src/main/resources/templates/searchUser.html
#	src/main/resources/templates/user.html

+++ src/main/java/com/sysc4806/project/models/Product.java
@@ -135,9 +135,9 @@ public class Product{
      * Getter method for the product's Average rating
      * @return - products average rating
      */
-    public int getAverageRating() {
+    public float getAverageRating() {
         if(reviews.isEmpty()) return 0;
-        return reviews.stream().mapToInt(Review::getRating).sum() / reviews.size();
+        return reviews.stream().mapToInt(Review::getRating).sum() / (float) reviews.size();
     }
 
     /**
